Regex transformer is able to do regex transformations
This change introduces a simple version of the transformer
that reads all the file in memory, does the transformation
using java regex and write the result to the file if it is
modified.

Java groups are allowed. In the form of (group) and $N in the
replacement. Where N = the group number.

IOException is propagated because I think it is the safest way
to do it for now. If we detect that we cannot modify a file
we should fail with IO up to Main program. Next CLs probably
will refine this more and introduce Fatal/Environment errors
vs user errors. But I think this is fine for now.

Change-Id: Id26fb6029e94e021ebef8869708c1506baab14ea