Make Reference object and generic Origins
This is what Matt and I discussed in a previous change, so that we
can make the references type safe.

The most contentious aspect of this change is that Reference is generic.
While this makes a bit more complex the changes method it has the
benefit that it can ensure that only changes generated by Origin are
used.

The alternative to not use generics is to add changes to the
Reference interface and set the fromRef parameter as an String.
This limits the usage of the method and puts the String -> Reference
logic in two methods. So I think the current approach is better.
'List of Changes of GitOrigin' doesn't read that bad.

Change-Id: I9fe46e763480f169752be04653ba181d412c0486