Allow specifying author line in {Git,Gerrit}Destination
Because we are not planning to duplicate commit history between
repositories, we will be merging commits when writing them out to Git,
so the author line will probably refer to some team, and shouldn't
refer to whoever happened to run Copybara.

So this change allows specifying the Author in the configuration for
the destination. It could be argued that it would make more sense to
specify this in the origin, since the origin knows who wrote its code,
but then the origin author info would have to be abstract (name +
e-mail + other things that the destination wants to know). So to keep
this simple we will specify the author on the destination.

It could be argued that the author timestamp should be specified based
on the origin version information, but that is out-of-scope for this
change.

This changes moves GitDestination.AbstractYaml to a top-level class
since it is no longer trivial.

Change-Id: I8595991d0c1e355ecc9dfa906ada3c1342ba7600