Remove workdir usage by origin, destination and transform
This change removes the availability of the workdir as an option to
those classes since its usage would be incorrect. For example a
workflow could decide to run several times on different workdirs.
None of those classes should make any assumption on what directory
will be run until their corresponding process method is called.

We still keep a fileSystem attribute since the folder destination
needs it for resolving the local-folder path. But we might aswell
remove it from the general options in the future.

Change-Id: Ifc41f20bbca7738b31a4a8f1c8083170556b7994