Make !Sequence transformation available to config
This change introduces the new !Sequence YAML element that can be used
to group a list of transformations in one element that can be reverted.

Removed ReversibleTransformation. It turns out that it was a bad idea.
As matvore suggested, changed to a method isReversible() in the Yaml
class so that we can get better errors.

Migrated Sequence tests to use Sequence.Yaml.

Added an sh integration test that uses Sequence + reverse workflow +
YAML anchor references to check that everything works.

Change-Id: I0cd104e3575f174605ed9d61cd8900c67dd57e15