!MoveFiles: Allow moving a subdirectory to the workdir root
This is to allow the internal GooglesMonoRepoSource (MRS) to pull
files from a single or multiple trees into the repo dir without moving
them.

Most users will use MRS with a repo dir. The MRS could by default move
all these to the workdir implicitly, but this will not be aplicable to
the users which actually must pull in multiple repo dirs.

Change-Id: Ida2e7f6418f3bf7fb437b6d1ebf34f6690e719a7