MoveFiles transform: allow moving from root to subdir
This allows moving all files in the workdir to a subdirectory. It also
has tweaks which will make it possible to move files from any
directory to a subdirectory of itself.

This includes a check to make sure the destination does not already
contain files in the case of moving to a subdirectory, since this is
kind of a mistake and would be hard to make work right if it were
intended.

Change-Id: Ie89756933608b096e46108ebf246c56b22dcb2e3