Support multiline Replace
This allows replacements that span more than one line. It also sets
the MULTILINE flag, which causes ^ and $ to match beginning and end of
lines. This is not strictly necessary, since \A and \Z can always be
used to match the beginning and end of lines, but some users will be
surprised if we don't actually turn on "multiline" mode in the regex
engine.

Change-Id: I614e312a0f47b503e2f7c113d86c464ce4c1f5ac