Be able to show the version of the binary
This change introduces the ability to extend Copybara to
show the version of the binary and other metadata. The
current implementation doesn't show any information since
we don't have public releases.

This change records the version always in the log so that
it is easier to debug issues. We also include the version
in the help.

Next: An internal change will use this to include the version
in the release among other build information.

Change-Id: I73296d7824fdd64b87f25e79efb7901b44994add