Make Workflow implementation code not inherit
Workflows inheriting made the differences between the workflows less
clear, and required a lot of code repetition in construction. It also
required repeating some subtle code, which was error prone.

Logging and console messages may be slightly different as a result of
code moving around. I removed a check that makes sure commit messages
end with "\n" in ITERATIVE mode, since the check seemed out-of-place
and not particularly helpful.

Change-Id: I78a4b2f9458099f5cd311817a17ef3572d6ec4a4