Allow destinations to use non HEAD baseline
This change introduces the ability to apply a worktree which its parent
is not up-to-date to HEAD in destination. The strategy we use is to
sync first to the parent baseline, apply the worktree and then try to
rebase the change to HEAD.

For git (And should be done for any other destination), merge conflicts
during the rebase are detected.

This is the first change on a series of changes to allow importing a
change from a Origin that is not up-to-date to destination HEAD.

Change-Id: Ia9ea21e4e12b35248f9a62759c48ac95b524783f