Move askForConfirmation to TransformResult
This is needed for the Skylark work since origins are not going
to be inside the workflow context during creation anymore (you
could create a destination and assign it multiple times). The
two alternatives considered are:

  - Use a wrapper object for each destination like Foo.Yaml ones.
    I want to get rid of them. It is a pain to maintain and we
    are already going to have our own Skylark method for creation.

  - Move askForConfirmation to each destination (That needs it).
    This would be good because we wouldn't have the weirdness of
    FolderDestination warning. But OTOH it prevents you to use
    the same destination twice in two modes. Besides, we decided
    that this flag was going to be in the workflow, and nothing
    appart of the language has changed. So I think this decision
    shouldn't change

The main con to this is that the validation for the flag is moved
to the "runtime". But AFAIK there is only one case were we enforce
this (Internal code). And in this case we should ignore the flag
and ask always (If we don't do already).

Change-Id: I7ed93fb20d844212e6dd66ce93e1feb4dd6ece3a