Add glob() Skylark function for PathMatcherBuilder
This change makes PathMatcherBuilder available to Skylark by adding a
new native function called 'glob'. This function is not exposed with
the namespace prefix 'core' so that it is easier and less verbose to
use. It also makes it more compatible with BUILD/Skylark files.

As part of this change I also introduce a functionality requested by
some users: To be able to exclude some paths from the included set. So
for example a call like:

  glob(include=['**/*.java'], exclude=['foo/Some.java'])

Includes all java files in subdirectories except for foo/Some.java.

Note that 'exclude' functionality is only visible to Skylark, not to
YAML configs.

As part of this change I also added the excludedInOrigin/Destination
fields in workflow. Initially I tried setting the default value to
'glob()' but it fails because of a cycle dependency on Skylark loading.
I changed to None as it seems more or less the same from the user POV.

Change-Id: I8aa06356b0a4650dd74b6971d414b084ce146dd2