Implement validation for Skylark configs
This change introduces a comparator to validate Skylark loaded configs
to the corresponding YAML files. I am assuming we are missing some
equals implementations. But lets add them on demand, once we have a
working origin to test with.

Sadly we cannot add testing since it requires a valid Origin. I'll
will add a test that checks that we show an error if the configs are
different once we have GitOrigin.

Change-Id: I3e8241ea3c6af1ed0e68575c049b5aeaeee48c48