Map all Workflow fields for Skylark
Map the remaining field that we were setting with constant values.

As part of this change I also made all the parameters of Workflow to
not be able to be used without a name (core.worflow(foo,bar,baz)).
This is terrible for readabily of the config, given the number of
parameters and also makes impossible to have some fields noneable
and some not. It also protect us from reordering the params (specially
dangerous given that we have many booleans).

Change-Id: I9f7aa1b4e1bbce1e25151c6cbd09214c55adf748