Add a hidden flag to git.origin for hook execution
This flags allows to execute a hook after the checkout happens.
This should not be generally used as we want to avoid shelling
out for transformations. This allows running gclient before
passing the control back to Copybara, so that we can get all
the code dependencies for the migration.

Hidden from documentation on purpose. If we see more gclient
usages we can think about adding more support for this.

Change-Id: I4a8e5b59f44b8a6a7460a1f9eb9896b2b88e82a7